Advancements, Dyes, Shields, Fixes and Tweaks
by MrNorwood
Added VBO model rendering

Fixed Armor positioning

by Alcater:
Added chemical dyes
Added new Dud versions
Added info to mechanists circle
Added burner press preheater
Added hbm command info
Added meteor and jungle dungeon to spawn command
Added ntm shields
Added material based shredding
Added more progression advancements
Added gas centrifuge upgrades
Added missing hazards to bedrock ores
Added 16x wire mold
Added low and high density elements
Added bedrock ore ore dict entries
Added shift click for various inventories
Added quark and inventory tweaks sorting to crates
Added new cadmium recipe
Added range info to fatman mininukes
Added press stamp rendering
Added asbestos to mining ship loot

Changed nuke drop to only happen at >15 radius
Changed nuke to ignore water when filling anyway
Changed nuke shockwave speed
Changed nuke dmg model
Changed glowing one dmg
Changed crystallizer jei
Changed bedrock ore balance
Changed anvil recipes
Changed fusion recipes to give more dusts
Changed battery i/o rate to make them more useable
Changed item frames and iron golems to be rad immune

Fixed satellite structure spawn
Fixed conveyor inventory interactions
Fixed assembler/chemplant inventory interactions
Fixed snow decontamination
Fixed boss advancements
Fixed scrap crafting
Fixed mechanists circle spawn room
Fixed volcano particle crash
Fixed geysir particle crash
Fixed gas explosion recursion crash
Fixed sliding blast door keypad mining
Fixed force field mining drops
Fixed various recipe loops and missing ones
Fixed small nuclear reactor i/o
Fixed pile breeding progress saving
Fixed mixer energy exploit
Fixed boiler energy exploit

Removed old rad tanks
Removed unused watz pellets

+++ src/main/java/com/hbm/render/amlfrom1710/Vec3.java
@@ -191,7 +191,7 @@ public class Vec3
      */
     public double length()
     {
-        return (double)MathHelper.sqrt(this.xCoord * this.xCoord + this.yCoord * this.yCoord + this.zCoord * this.zCoord);
+        return MathHelper.sqrt(this.xCoord * this.xCoord + this.yCoord * this.yCoord + this.zCoord * this.zCoord);
     }
     
     public double lengthSquared(){
